// infrastructure/mappers/movie.mapper.ts
import { GetMovieInterface } from "../interfaces/get-moviedb.interface";
import { CompleteMovie, Movie } from "../interfaces/movie.interface";
import { Result } from "../interfaces/moviedb-response";

export class MovieMapper {
    static fromTheMovieDBToMovie = (movie: Result): Movie => {
        return {
            id: movie.id,
            title: movie.title,
            description: movie.overview,
            releaseDate: new Date(movie.release_date),
            rating: movie.vote_average,
            poster: `https://image.tmdb.org/t/p/w500${movie.poster_path}`,
            backdrop: `https://image.tmdb.org/t/p/w500${movie.backdrop_path}`
        }
    }

     static fromTheMovieDBToMovieGet = (movie: GetMovieInterface): CompleteMovie => {
        return {
            id: movie.id,
            title: movie.title,
            description: movie.overview,
            releaseDate: new Date(movie.release_date),
            rating: movie.vote_average,
            poster: `https://image.tmdb.org/t/p/w500${movie.poster_path}`,
            backdrop: `https://image.tmdb.org/t/p/w500${movie.backdrop_path}`,
            budget: movie.budget,
            duration: movie.runtime,
            genres: movie.genres.map(g => g.name),
            originalTitle: movie.original_title,
            productionCompanies: movie.production_companies.map( c => c.name)
        }
    }

    // static fromTheMovieBDToCompleteMovie = (movie: MovieDBResponse): CompleteMovie => {
    //     return(
    //          id: movie.id,
    //         title: movie.title,
    //         description: movie.overview,
    //         releaseDate: new Date(movie.release_date),
    //         rating: movie.vote_average,
    //         poster: `https://image.tmdb.org/t/p/w500${movie.poster_path}`,
    //         backdrop: `https://image.tmdb.org/t/p/w500${movie.backdrop_path}`
    //     )
    // }
}